pyproject.toml
[tool.poetry]
name = "event-api"
version = "0.1.0"
description = ""
authors = ["Fagner Pinheiro <fagnerpinheirosantos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.3.6"
redis = "^5.0.3"
drf-yasg = "^1.21.7"
setuptools = "^69.2.0"
django-environ = "^0.11.2"
dj-database-url = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
